refactor: front controller 패턴 적용
